Resolve calls in generated asm (bpf-profile #7)
