+++ src/gen/asm.rs
@@ -10,6 +10,7 @@ pub struct Instruction {
 }
 
 use super::{Error, Result};
+use crate::config::Address;
 use lazy_static::lazy_static;
 use regex::Regex;
 
@@ -56,6 +57,21 @@ impl Instruction {
     pub fn is_exit(&self) -> bool {
         self.text == "exit"
     }
+
+    /// Returns address of a call target or error if instruction is not a call.
+    pub fn extract_call_target(&self, lc: usize) -> Result<Address> {
+        if !self.is_call() {
+            return Err(Error::TraceNotCall(self.text(), lc));
+        }
+        let mut pair = self.text.split_whitespace(); // => "call something"
+        let _ = pair
+            .next()
+            .ok_or_else(|| Error::TraceParsing(self.text(), lc))?;
+        let address = pair
+            .next()
+            .ok_or_else(|| Error::TraceParsing(self.text(), lc))?;
+        Ok(hex_str_to_address(address))
+    }
 }
 
 use std::fmt;
@@ -76,6 +92,7 @@ pub struct Source {
     ixs: Vec<Instruction>,
 }
 
+use super::dump::Resolver;
 use std::io::Write;
 
 impl Source {
@@ -101,16 +118,32 @@ impl Source {
     }
 
     /// Writes all lines of the listing to a file.
-    pub fn write(&self, mut output: impl Write) -> Result<()> {
+    pub fn write(&self, mut output: impl Write, resolver: &Resolver) -> Result<()> {
         for i in 0..self.ixs.len() {
             let ix = &self.ixs[i];
             if i == 0 && ix.is_empty() {
                 writeln!(output, "Generated BPF assembly code for QCacheGrind")?;
-            } else {
-                writeln!(output, "{}", ix)?;
+                continue;
             }
+            if ix.is_call() {
+                let address = ix.extract_call_target(i)?;
+                let name = resolver.resolve(address);
+                let ix = Instruction {
+                    pc: ix.pc(),
+                    text: format!("call {}", &name),
+                };
+                writeln!(output, "{}", ix)?;
+                continue;
+            }
+            writeln!(output, "{}", ix)?;
         }
         output.flush()?;
         Ok(())
     }
+}
+
+/// Converts a hex number string representation to integer Address.
+fn hex_str_to_address(s: &str) -> Address {
+    let a = s.trim_start_matches("0x");
+    Address::from_str_radix(a, 16).expect("Invalid address")
 }
