+++ src/gen/trace.rs
@@ -1,15 +1,12 @@
-//! bpf-profile implementation of the profile struct.
+//! bpf-profile trace module.
+//! Implements parsing of the trace file and generating the profile.
 
 use super::{fileutil, Error, Result};
-use crate::config::{Address, Cost, Map, ProgramCounter, GROUND_ZERO};
+use crate::config::{Cost, Map, ProgramCounter};
 use lazy_static::lazy_static;
 use regex::Regex;
-use std::collections::BTreeMap;
 use std::io::{BufRead, BufReader, Write};
 use std::path::Path;
-
-type Functions = Map<Address, Function>;
-type Costs = BTreeMap<ProgramCounter, Cost>; // sort by pc
 
 /// Checks the trace file contains expected header line.
 pub fn contains_standard_header(mut reader: impl BufRead) -> Result<bool> {
@@ -33,6 +30,8 @@ pub fn contains_standard_header(mut reader: impl BufRead) -> Result<bool> {
 
 use super::asm::{self, Instruction};
 use super::dump::{self, Resolver};
+use super::profile::{self, Call, Function, Functions};
+use crate::config::GROUND_ZERO;
 
 /// Represents the profile.
 #[derive(Debug)]
@@ -82,7 +81,7 @@ impl Profile {
     /// Writes the generated assembly file.
     pub fn write_asm(&self, asm_path: &Path) -> Result<()> {
         let output = fileutil::open_w(asm_path)?;
-        self.asm.write(output)
+        self.asm.write(output, &self.dump)
     }
 
     /// Writes the profile data in the callgrind file format.
@@ -94,7 +93,7 @@ impl Profile {
         writeln!(output, "events: Instructions")?;
         writeln!(output, "totals: {}", self.total_cost)?;
         writeln!(output, "fl={}", asm_fl)?;
-        write_callgrind_functions(&self.functions, output)?;
+        profile::write_callgrind_functions(output, &self.functions)?;
         Ok(())
     }
 
@@ -112,7 +111,7 @@ impl Profile {
 
     /// Adds next call to the call stack.
     fn push_call(&mut self, call: Call, first_pc: ProgramCounter) {
-        let address = call.address;
+        let address = call.address();
         tracing::debug!("Profile.push_call {}", address);
         self.ground.push_call(call);
         #[allow(clippy::map_entry)]
@@ -126,159 +125,13 @@ impl Profile {
     /// Removes finished call from the call stack and adds it to the caller.
     fn pop_call(&mut self) {
         let call = self.ground.pop_call();
-        tracing::debug!("Profile.pop_call {}", &call.address);
+        tracing::debug!("Profile.pop_call {}", &call.address());
         if !call.is_ground() {
             let f = self
                 .functions
-                .get_mut(&call.caller)
+                .get_mut(&call.caller())
                 .expect("Caller not found in registry of functions");
             f.add_call(call);
-        }
-    }
-}
-
-/// Represents a function which will be dumped into a profile.
-#[derive(Debug)]
-struct Function {
-    address: Address,
-    name: String,
-    costs: Costs,
-    calls: Vec<Call>,
-}
-
-impl Function {
-    /// Creates initial function object which stores total cost of entire program.
-    fn ground_zero() -> Self {
-        Function {
-            address: GROUND_ZERO,
-            name: "GROUND_ZERO".into(),
-            costs: BTreeMap::new(),
-            calls: Vec::new(),
-        }
-    }
-
-    /// Creates new function object.
-    fn new(address: Address, first_pc: ProgramCounter, dump: &mut Resolver) -> Self {
-        assert_ne!(address, GROUND_ZERO);
-        Function {
-            address,
-            name: dump.resolve(address, first_pc),
-            costs: BTreeMap::new(),
-            calls: Vec::new(),
-        }
-    }
-
-    /// Increments the immediate cost of the function.
-    fn increment_cost(&mut self, pc: ProgramCounter) {
-        tracing::debug!("Function({}).increment_cost", self.address);
-        let c = *self.costs.entry(pc).or_insert(0);
-        self.costs.insert(pc, c + 1);
-    }
-
-    /// Adds finished enclosed call for this function.
-    fn add_call(&mut self, call: Call) {
-        tracing::debug!("Function({}).add_call {}", self.address, call.address);
-        self.calls.push(call);
-    }
-}
-
-/// Represents a function call.
-#[derive(Clone, Debug)]
-struct Call {
-    address: Address,
-    caller: Address,
-    caller_pc: ProgramCounter,
-    cost: Cost,
-    callee: Box<Option<Call>>,
-    depth: usize,
-}
-
-impl Call {
-    /// Creates new call object.
-    fn new(address: Address, caller_pc: ProgramCounter) -> Self {
-        Call {
-            address,
-            caller: Address::default(),
-            caller_pc,
-            cost: 0,
-            callee: Box::new(None),
-            depth: 0,
-        }
-    }
-
-    /// Creates new call object from a trace instruction (which must be a call).
-    fn from(ix: &Instruction, lc: usize) -> Result<Self> {
-        let text = ix.text();
-        if !ix.is_call() {
-            return Err(Error::TraceNotCall(text));
-        }
-        let mut pair = text.split_whitespace(); // => "call something"
-        let _ = pair
-            .next()
-            .ok_or_else(|| Error::TraceParsing(ix.text(), lc))?;
-        let address = pair
-            .next()
-            .ok_or_else(|| Error::TraceParsing(ix.text(), lc))?;
-        Ok(Call::new(hex_str_to_address(address), ix.pc()))
-    }
-
-    /// Checks if the call is the root ("ground zero").
-    fn is_ground(&self) -> bool {
-        self.address == GROUND_ZERO
-    }
-
-    /// Increments the cost of this call.
-    fn increment_cost(&mut self, pc: ProgramCounter, functions: &mut Functions) {
-        tracing::debug!("Call({}).increment_cost", self.address);
-        match *self.callee {
-            Some(ref mut callee) => {
-                callee.increment_cost(pc, functions);
-            }
-            None => {
-                self.cost += 1;
-                let f = functions
-                    .get_mut(&self.address)
-                    .expect("Call address not found in the registry of functions");
-                f.increment_cost(pc);
-            }
-        }
-    }
-
-    /// Adds next call to the call stack.
-    fn push_call(&mut self, mut call: Call) {
-        tracing::debug!(
-            "Call({}).push_call {} depth={}",
-            self.address,
-            call.address,
-            self.depth
-        );
-        self.depth += 1;
-        match *self.callee {
-            Some(ref mut callee) => {
-                callee.push_call(call);
-            }
-            None => {
-                call.caller = self.address;
-                let old = std::mem::replace(&mut *self.callee, Some(call));
-                assert!(old.is_none());
-            }
-        }
-    }
-
-    /// Removes current call from the call stack.
-    fn pop_call(&mut self) -> Call {
-        tracing::debug!("Call({}).pop_call depth={}", self.address, self.depth);
-        if self.depth == 0 {
-            panic!("Exit without call");
-        }
-        self.depth -= 1;
-        let callee = self.callee.as_mut().as_mut().expect("Missing callee");
-        if callee.callee.is_some() {
-            callee.pop_call()
-        } else {
-            let call = self.callee.take().expect("Missing callee");
-            self.cost += call.cost;
-            call
         }
     }
 }
@@ -336,51 +189,11 @@ pub fn parse(mut reader: impl BufRead, prof: &mut Profile) -> Result<()> {
         // Keep here the last non-call line to process further
     }
 
-    if prof.ground.depth > 0 {
-        tracing::warn!("Unbalanced call/exit: {}", &prof.ground.depth);
-        for _ in 0..prof.ground.depth {
+    if prof.ground.depth() > 0 {
+        tracing::warn!("Unbalanced call/exit: {}", &prof.ground.depth());
+        for _ in 0..prof.ground.depth() {
             prof.pop_call();
         }
     }
     Ok(())
-}
-
-/// Writes information about calls of functions and their costs.
-fn write_callgrind_functions(functions: &Functions, mut output: impl Write) -> Result<()> {
-    let mut statistics = Map::new();
-
-    for (a, f) in functions {
-        if *a == GROUND_ZERO {
-            continue;
-        }
-
-        writeln!(output)?;
-        writeln!(output, "fn={}", f.name)?;
-        for (pc, cost) in &f.costs {
-            writeln!(output, "{} {}", pc, cost)?;
-        }
-
-        statistics.clear();
-        for c in &f.calls {
-            let key = (c.caller_pc, c.address);
-            let stat = statistics.entry(key).or_insert((0_usize, 0_usize));
-            let number_of_calls = stat.0 + 1;
-            let inclusive_cost = stat.1 + c.cost;
-            statistics.insert(key, (number_of_calls, inclusive_cost));
-        }
-        for ((pc, address), (number_of_calls, inclusive_cost)) in &statistics {
-            writeln!(output, "cfn={}", functions[address].name)?;
-            writeln!(output, "calls={} 0x{:x}", number_of_calls, address)?;
-            writeln!(output, "{} {}", pc, inclusive_cost)?;
-        }
-    }
-
-    output.flush()?;
-    Ok(())
-}
-
-/// Converts a hex number string representation to integer Address.
-fn hex_str_to_address(s: &str) -> Address {
-    let a = s.trim_start_matches("0x");
-    Address::from_str_radix(a, 16).expect("Invalid address")
 }
